# perf(jce): drop reflect in writer