--- a/binary/jce/structs.go
+++ b/binary/jce/structs.go
@@ -5,6 +5,7 @@ type IJceStruct interface {
 	ReadFrom(*JceReader)
 }
 
+//go:generate go run gen/structs_parser.go -f structs_tobytes.go
 type (
 	RequestPacket struct {
 		IVersion     int16             `jceId:"1"`
@@ -539,12 +540,6 @@ type (
 	}
 )
 
-func (pkt *RequestPacket) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
 func (pkt *RequestPacket) ReadFrom(r *JceReader) {
 	pkt.IVersion = r.ReadInt16(1)
 	pkt.CPacketType = r.ReadByte(2)
@@ -560,20 +555,8 @@ func (pkt *RequestPacket) ReadFrom(r *JceReader) {
 	pkt.Status = r.ReadMapStrStr(10)
 }
 
-func (pkt *RequestDataVersion3) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
 func (pkt *RequestDataVersion3) ReadFrom(r *JceReader) {
 	pkt.Map = r.ReadMapStrByte(0)
-}
-
-func (pkt *RequestDataVersion2) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
 }
 
 func (pkt *RequestDataVersion2) ReadFrom(r *JceReader) {
@@ -625,12 +608,6 @@ func (pkt *BigDataIPInfo) ReadFrom(r *JceReader) {
 	pkt.Port = r.ReadInt64(2)
 }
 
-func (pkt *SvcReqRegister) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
 func (pkt *SvcRespRegister) ReadFrom(r *JceReader) {
 	pkt.Uin = r.ReadInt64(0)
 	pkt.Bid = r.ReadInt64(1)
@@ -652,24 +629,6 @@ func (pkt *SvcRespRegister) ReadFrom(r *JceReader) {
 	pkt.ExtOnlineStatus = r.ReadInt64(17)
 }
 
-func (pkt *FriendListRequest) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
-func (pkt *SummaryCardReq) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
-func (pkt *SummaryCardReqSearch) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
 func (pkt *FriendInfo) ReadFrom(r *JceReader) {
 	pkt.FriendUin = r.ReadInt64(0)
 	pkt.GroupId = r.ReadByte(1)
@@ -683,12 +642,6 @@ func (pkt *FriendInfo) ReadFrom(r *JceReader) {
 	pkt.CardID = r.ReadBytes(41)
 }
 
-func (pkt *TroopListRequest) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
 func (pkt *TroopNumber) ReadFrom(r *JceReader) {
 	pkt.GroupUin = r.ReadInt64(0)
 	pkt.GroupCode = r.ReadInt64(1)
@@ -697,12 +650,6 @@ func (pkt *TroopNumber) ReadFrom(r *JceReader) {
 	pkt.MemberNum = r.ReadInt64(19)
 	pkt.GroupOwnerUin = r.ReadInt64(23)
 	pkt.MaxGroupMemberNum = r.ReadInt64(29)
-}
-
-func (pkt *TroopMemberListRequest) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
 }
 
 func (pkt *TroopMemberInfo) ReadFrom(r *JceReader) {
@@ -799,34 +746,4 @@ func (pkt *InstanceInfo) ReadFrom(r *JceReader) {
 	pkt.Platform = r.ReadInt64(2)
 	pkt.ProductType = r.ReadInt64(3)
 	pkt.ClientType = r.ReadInt64(4)
-}
-
-func (pkt *SvcRespPushMsg) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
-func (pkt *ModifyGroupCardRequest) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
-func (pkt *SvcReqGetDevLoginInfo) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
-func (pkt *SvcReqRegisterNew) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
-}
-
-func (pkt *DelFriendReq) ToBytes() []byte {
-	w := NewJceWriter()
-	w.WriteJceStructRaw(pkt)
-	return w.Bytes()
 }