--- a/binary/jce/structs.go
+++ b/binary/jce/structs.go
@@ -1,7 +1,7 @@
 package jce
 
 type IJceStruct interface {
-	// ToBytes() []byte
+	ToBytes() []byte
 	ReadFrom(*JceReader)
 }