--- a/binary/jce/writer.go
+++ b/binary/jce/writer.go
@@ -4,9 +4,6 @@ import (
 	"bytes"
 	goBinary "encoding/binary"
 	"math"
-	"reflect"
-	"strconv"
-	"sync"
 )
 
 type JceWriter struct {
@@ -166,68 +163,6 @@ func (w *JceWriter) WriteBytesSlice(l [][]byte, tag byte) {
 	}
 }
 
-func (w *JceWriter) WriteSlice(i interface{}, tag byte) {
-	va := reflect.ValueOf(i)
-	if va.Kind() != reflect.Slice {
-		panic("JceWriter.WriteSlice: not a slice")
-	}
-	w.writeSlice(va, tag)
-}
-
-func (w *JceWriter) writeSlice(slice reflect.Value, tag byte) {
-	if slice.Kind() != reflect.Slice {
-		return
-	}
-	w.writeHead(9, tag)
-	if slice.Len() == 0 {
-		w.writeHead(12, 0) // w.WriteInt32(0, 0)
-		return
-	}
-	w.WriteInt32(int32(slice.Len()), 0)
-	for i := 0; i < slice.Len(); i++ {
-		v := slice.Index(i)
-		w.writeObject(v, 0)
-	}
-}
-
-func (w *JceWriter) WriteJceStructSlice(l []IJceStruct, tag byte) {
-	w.writeHead(9, tag)
-	if len(l) == 0 {
-		w.writeHead(12, 0) // w.WriteInt32(0, 0)
-		return
-	}
-	w.WriteInt32(int32(len(l)), 0)
-	for _, v := range l {
-		w.WriteJceStruct(v, 0)
-	}
-}
-
-func (w *JceWriter) WriteMap(m interface{}, tag byte) {
-	va := reflect.ValueOf(m)
-	if va.Kind() != reflect.Map {
-		panic("JceWriter.WriteMap: not a map")
-	}
-	w.writeMap(va, tag)
-}
-
-func (w *JceWriter) writeMap(m reflect.Value, tag byte) {
-	if m.IsNil() {
-		w.writeHead(8, tag)
-		w.writeHead(12, 0) // w.WriteInt32(0, 0)
-		return
-	}
-	if m.Kind() != reflect.Map {
-		return
-	}
-	w.writeHead(8, tag)
-	w.WriteInt32(int32(m.Len()), 0)
-	iter := m.MapRange()
-	for iter.Next() {
-		w.writeObject(iter.Key(), 0)
-		w.writeObject(iter.Value(), 1)
-	}
-}
-
 func (w *JceWriter) writeMapStrStr(m map[string]string, tag byte) {
 	if m == nil {
 		w.writeHead(8, tag)
@@ -268,105 +203,6 @@ func (w *JceWriter) writeMapStrMapStrBytes(m map[string]map[string][]byte, tag b
 		w.WriteString(k, 0)
 		w.writeMapStrBytes(v, 1)
 	}
-}
-
-func (w *JceWriter) writeObject(v reflect.Value, tag byte) {
-	k := v.Kind()
-	if k == reflect.Map {
-		switch o := v.Interface().(type) {
-		case map[string]string:
-			w.writeMapStrStr(o, tag)
-		case map[string][]byte:
-			w.writeMapStrBytes(o, tag)
-		case map[string]map[string][]byte:
-			w.writeMapStrMapStrBytes(o, tag)
-		default:
-			w.writeMap(v, tag)
-		}
-		return
-	}
-	if k == reflect.Slice {
-		switch o := v.Interface().(type) {
-		case []byte:
-			w.WriteBytes(o, tag)
-		case []IJceStruct:
-			w.WriteJceStructSlice(o, tag)
-		default:
-			w.writeSlice(v, tag)
-		}
-		return
-	}
-	switch k {
-	case reflect.Uint8, reflect.Int8:
-		w.WriteByte(*(*byte)(pointerOf(v)), tag)
-	case reflect.Uint16, reflect.Int16:
-		w.WriteInt16(*(*int16)(pointerOf(v)), tag)
-	case reflect.Uint32, reflect.Int32:
-		w.WriteInt32(*(*int32)(pointerOf(v)), tag)
-	case reflect.Uint64, reflect.Int64:
-		w.WriteInt64(*(*int64)(pointerOf(v)), tag)
-	case reflect.String:
-		w.WriteString(v.String(), tag)
-	default:
-		switch o := v.Interface().(type) {
-		case IJceStruct:
-			w.WriteJceStruct(o, tag)
-		case float32:
-			w.WriteFloat32(o, tag)
-		case float64:
-			w.WriteFloat64(o, tag)
-		}
-	}
-}
-
-type decoder struct {
-	index int
-	id    int
-}
-
-var decoderCache = sync.Map{}
-
-// WriteJceStructRaw 写入 Jce 结构体
-func (w *JceWriter) WriteJceStructRaw(s interface{}) {
-	t := reflect.TypeOf(s)
-	if t.Kind() != reflect.Ptr {
-		return
-	}
-	t = t.Elem()
-	v := reflect.ValueOf(s).Elem()
-	var jceDec []decoder
-	dec, ok := decoderCache.Load(t)
-	if ok { // 从缓存中加载
-		jceDec = dec.([]decoder)
-	} else { // 初次反射
-		jceDec = make([]decoder, 0, t.NumField())
-		for i := 0; i < t.NumField(); i++ {
-			field := t.Field(i)
-			strId := field.Tag.Get("jceId")
-			if strId == "" {
-				continue
-			}
-			id, err := strconv.Atoi(strId)
-			if err != nil {
-				continue
-			}
-			jceDec = append(jceDec, decoder{
-				index: i,
-				id:    id,
-			})
-		}
-		decoderCache.Store(t, jceDec) // 存入缓存
-	}
-	for _, dec := range jceDec {
-		obj := v.Field(dec.index)
-		w.writeObject(obj, byte(dec.id))
-	}
-}
-
-func (w *JceWriter) WriteJceStruct(s IJceStruct, tag byte) {
-	w.writeHead(10, tag)
-	w.WriteJceStructRaw(s)
-	w.writeHead(11, 0)
 }
 
 func (w *JceWriter) Bytes() []byte {