… binary/jce/gen/structs_parser_tmp.go
… binary/jce/reflectx.go
… binary/jce/structs_tobytes.go
… binary/jce/writer_test.go