@wdvxdr1123 2021-12-10
generator有点诡异了，我觉得应该用 `go/parser` 之类的