@fumiama 2021-12-10
> generator有点诡异了，我觉得应该用 `go/parser` 之类的

确实如此，不过刚才试了一下，`go/parser`要麻烦很多倍，在效率上是得不偿失的。