@fumiama 2021-12-10
我修改了一下generator，目前这样并没有写死为const，也便于修改代码以及查错，应该还可以接受。